move files config and constants to shared workspace
