+++ web/lib/machines/drop.ts
@@ -1,7 +1,6 @@
-import { DropEventType, DropState } from '../constants';
+import type { AnyDropEvent, DropContext } from 'types/drop';
+import { DropEventType, DropState } from '@shared/lib/constants';
 import { createMachine, TransitionsConfig } from 'xstate';
-import type { AnyDropEvent } from 'types/drop';
-import type { DropContext } from 'types/drop';
 import { raise as baseRaise } from 'xstate/lib/actions';
 
 export const initDropContext = (): DropContext => ({
