+++ web/lib/machines/grab.ts
@@ -1,7 +1,6 @@
-import { GrabEventType, GrabState } from '../constants';
+import type { AnyGrabEvent, GrabContext } from 'types/grab';
+import { GrabEventType, GrabState } from '@shared/lib/constants';
 import { createMachine, TransitionsConfig } from 'xstate';
-import type { AnyGrabEvent } from 'types/grab';
-import type { GrabContext } from 'types/grab';
 import { raise as baseRaise } from 'xstate/lib/actions';
 
 export const initGrabContext = (): GrabContext => ({
