+++ web/molecules/steps/SecretInputCard.tsx
@@ -12,7 +12,7 @@ import StepCard from './StepCard';
 import { useDropContext } from 'contexts/DropContext';
 import type { PayloadInputMode } from 'types/common';
 import { Captcha } from 'atoms/Captcha';
-import { ACCEPTED_FILE_TYPES, MAX_PAYLOAD_SIZE } from 'config/files';
+import { ACCEPTED_FILE_TYPES, MAX_PAYLOAD_SIZE } from '@shared/config/files';
 
 export const SecretInputCard = () => {
     const [mode, setMode] = useState<PayloadInputMode>('text');
