+++ web/types/grab.ts
@@ -1,5 +1,5 @@
 import type { BaseContext } from './common';
-import { GrabEventType } from 'lib/constants';
+import { GrabEventType } from '@shared/lib/constants';
 import type { EventObject } from 'xstate/lib/types';
 import type Peer from 'peerjs';
 import { useGrab } from 'hooks/use-grab';
